server/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
crewai==0.11.0
pyautogen==0.2.2
psutil==5.9.8
python-multipart==0.0.9
pynvml==11.5.0
requests==2.31.0
typing-extensions==4.9.0 